NeoVintageous fix
If Sublime Text is in the NeoVintageous command mode, the insert
position has to be shifted in some situations. This depends on the
characters below and to the right of the current cursor position. A
shift is necessary if the next character is a whitespace (space, tab,
the end of a line) or the cursor is at the end of the file. Additionally
it has to be ensured that the cursor is not on a punctuation character,
because the text should be inserted before in almost all cases.

This should work in all reasonable situations. Note, however, that some
of the characters allowed in BibTeX entry keys will break this.
Nevertheless it can be expected that these occur only rarely.